send notification for transaction item actions

+++ app/Traits/TransactionAccessors.php
@@ -96,7 +96,7 @@ trait TransactionAccessors
                         if ($this->pad->hasPrices()) {
                             $data['quantity'] = number_format($data['quantity'], 2, thousands_separator:'');
                             $data['unit_price'] = number_format($data['unit_price'], 2, thousands_separator:'');
-                            $data['discount'] = $data['discount'] ?? 0.00;
+                            $data['discount'] = (float) ($data['discount'] ?? 0.00);
 
                             $unitPrice = userCompany()->isPriceBeforeVAT() ? $data['unit_price'] : number_format($data['unit_price'] / 1.15, 2, thousands_separator:'');
                             $data['total'] = number_format($unitPrice * $data['quantity'], 2, thousands_separator:'');
